Add release image validation for NodePools
